Fix warnings

+++ src/Plugin.cpp
@@ -203,7 +203,7 @@ void Plugin::update() {
       return a->distanceToCamera() < b->distanceToCamera();
     });
 
-    for (size_t i = 0; i < sortedLabels.size(); ++i) {
+    for (int i = 0; i < static_cast<int>(sortedLabels.size()); ++i) {
       // a little bit hacky... It probably breaks, when more than 100 labels are present.
       sortedLabels[i]->setSortKey(static_cast<int>(cs::utils::DrawOrder::eTransparentItems) - i);
     }
